Flytt tilRestYtelsePerioder inn i TilkjentYtelseUtils
