+++ src/main/kotlin/no/nav/familie/ba/sak/ekstern/restDomene/RestTilkjentYtelse.kt
@@ -3,7 +3,7 @@ package no.nav.familie.ba.sak.ekstern.restDomene
 import no.nav.familie.ba.sak.common.toYearMonth
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
-import no.nav.familie.ba.sak.kjerne.beregning.domene.tilRestYtelsePerioder
+import no.nav.familie.ba.sak.kjerne.beregning.tilRestYtelsePerioder
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
 import java.time.LocalDate
 import java.time.YearMonth
