+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/TilkjentYtelseUtils.kt
@@ -7,11 +7,13 @@ import no.nav.familie.ba.sak.common.erDagenFør
 import no.nav.familie.ba.sak.common.førsteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.inkluderer
 import no.nav.familie.ba.sak.common.sisteDagIInneværendeMåned
+import no.nav.familie.ba.sak.ekstern.restDomene.RestYtelsePeriode
 import no.nav.familie.ba.sak.kjerne.beregning.UtvidetBarnetrygdUtil.finnUtvidetVilkår
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.EndretUtbetalingAndelMedAndelerTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.InternPeriodeOvergangsstønad
+import no.nav.familie.ba.sak.kjerne.beregning.domene.RestYtelsePeriodeUtenDatoer
 import no.nav.familie.ba.sak.kjerne.beregning.domene.TilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
 import no.nav.familie.ba.sak.kjerne.beregning.domene.medEndring
@@ -21,6 +23,10 @@ import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonType
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersonopplysningGrunnlag
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
+import no.nav.familie.ba.sak.kjerne.tidslinje.komposisjon.slåSammenLike
+import no.nav.familie.ba.sak.kjerne.tidslinje.månedPeriodeAv
+import no.nav.familie.ba.sak.kjerne.tidslinje.tidspunkt.tilYearMonth
+import no.nav.familie.ba.sak.kjerne.tidslinje.tilTidslinje
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Vilkårsvurdering
 import java.math.BigDecimal
 import java.time.LocalDate
@@ -320,3 +326,37 @@ internal data class BeregnetAndel(
     val sats: Int,
     val prosent: BigDecimal,
 )
+
+fun List<AndelTilkjentYtelse>.tilRestYtelsePerioder(): List<RestYtelsePeriode> {
+    val restYtelsePeriodeTidslinjePerAktørOgTypeSlåttSammen =
+        this.groupBy { Pair(it.aktør, it.type) }
+            .mapValues { (_, andelerTilkjentYtelse) ->
+                andelerTilkjentYtelse
+                    .tilRestYtelsePeriodeUtenDatoerTidslinje()
+                    .slåSammenLike()
+            }
+
+    return restYtelsePeriodeTidslinjePerAktørOgTypeSlåttSammen
+        .flatMap { (_, andelerTidslinje) -> andelerTidslinje.perioder() }
+        .mapNotNull { periode ->
+            periode.innhold?.let { innhold ->
+                RestYtelsePeriode(
+                    beløp = innhold.kalkulertUtbetalingsbeløp,
+                    stønadFom = periode.fraOgMed.tilYearMonth(),
+                    stønadTom = periode.tilOgMed.tilYearMonth(),
+                    ytelseType = innhold.ytelseType,
+                    skalUtbetales = innhold.skalUtbetales,
+                )
+            }
+        }
+}
+
+private fun List<AndelTilkjentYtelse>.tilRestYtelsePeriodeUtenDatoerTidslinje() =
+    this.map {
+        månedPeriodeAv(
+            fraOgMed = it.stønadFom,
+            tilOgMed = it.stønadTom,
+            innhold = RestYtelsePeriodeUtenDatoer(kalkulertUtbetalingsbeløp = it.kalkulertUtbetalingsbeløp, it.type, it.prosent > BigDecimal.ZERO),
+        )
+    }
+        .tilTidslinje()
