+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/domene/AndelTilkjentYtelse.kt
@@ -22,16 +22,11 @@ import no.nav.familie.ba.sak.common.YearMonthConverter
 import no.nav.familie.ba.sak.common.førsteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.sisteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.toYearMonth
-import no.nav.familie.ba.sak.ekstern.restDomene.RestYtelsePeriode
 import no.nav.familie.ba.sak.integrasjoner.økonomi.YtelsetypeBA
 import no.nav.familie.ba.sak.kjerne.beregning.AndelTilkjentYtelseForVedtaksperioderTidslinje
 import no.nav.familie.ba.sak.kjerne.beregning.AndelTilkjentYtelseTidslinje
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Person
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
-import no.nav.familie.ba.sak.kjerne.tidslinje.komposisjon.slåSammenLike
-import no.nav.familie.ba.sak.kjerne.tidslinje.månedPeriodeAv
-import no.nav.familie.ba.sak.kjerne.tidslinje.tidspunkt.tilYearMonth
-import no.nav.familie.ba.sak.kjerne.tidslinje.tilTidslinje
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.utledSegmenter
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.PersonResultat
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.domene.Regelverk
@@ -204,40 +199,6 @@ data class RestYtelsePeriodeUtenDatoer(
     val ytelseType: YtelseType,
     val skalUtbetales: Boolean,
 )
-
-fun List<AndelTilkjentYtelse>.tilRestYtelsePerioder(): List<RestYtelsePeriode> {
-    val restYtelsePeriodeTidslinjePerAktørOgTypeSlåttSammen =
-        this.groupBy { Pair(it.aktør, it.type) }
-            .mapValues { (_, andelerTilkjentYtelse) ->
-                andelerTilkjentYtelse
-                    .tilRestYtelsePeriodeUtenDatoerTidslinje()
-                    .slåSammenLike()
-            }
-
-    return restYtelsePeriodeTidslinjePerAktørOgTypeSlåttSammen
-        .flatMap { (_, andelerTidslinje) -> andelerTidslinje.perioder() }
-        .mapNotNull { periode ->
-            periode.innhold?.let { innhold ->
-                RestYtelsePeriode(
-                    beløp = innhold.kalkulertUtbetalingsbeløp,
-                    stønadFom = periode.fraOgMed.tilYearMonth(),
-                    stønadTom = periode.tilOgMed.tilYearMonth(),
-                    ytelseType = innhold.ytelseType,
-                    skalUtbetales = innhold.skalUtbetales,
-                )
-            }
-        }
-}
-
-private fun List<AndelTilkjentYtelse>.tilRestYtelsePeriodeUtenDatoerTidslinje() =
-    this.map {
-        månedPeriodeAv(
-            fraOgMed = it.stønadFom,
-            tilOgMed = it.stønadTom,
-            innhold = RestYtelsePeriodeUtenDatoer(kalkulertUtbetalingsbeløp = it.kalkulertUtbetalingsbeløp, it.type, it.prosent > BigDecimal.ZERO),
-        )
-    }
-        .tilTidslinje()
 
 fun List<AndelTilkjentYtelseMedEndreteUtbetalinger>.lagVertikaleSegmenter(): Map<LocalDateSegment<Int>, List<AndelTilkjentYtelseMedEndreteUtbetalinger>> {
     return this.utledSegmenter()
